% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vf_locations.R
\name{add_vf_locations}
\alias{add_vf_locations}
\title{Add VF Locations}
\usage{
add_vf_locations(data, type = "24-2")
}
\arguments{
\item{data}{}
}
\value{
data with numbered locations
}
\description{
This function right joins the data to vf_map, adding all locations
 and introducing NA for areas with no data (which are supposed to be empty)
}
\examples{
add_vf_locations(data)
}
