% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_empty.R
\name{strip_empty}
\alias{strip_empty}
\title{Strip intentionally-placed empty grobs from facet plot}
\usage{
strip_empty(p)
}
\arguments{
\item{p}{}
}
\description{
Strip intentionally-placed empty grobs from facet plot
}
\examples{

# Strip empty facets
g <- strip_empty(p)

# Next, plot grob object
grid::grid.newpage()
grid:: grid.draw(g)
}
