% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{theme_clean}
\alias{theme_clean}
\title{Theme for clean plots
Use for 24-2 heatmaps, histograms, etc.}
\usage{
theme_clean(...)
}
\description{
Theme for clean plots
Use for 24-2 heatmaps, histograms, etc.
}
\examples{
p + theme_clean()
}
