% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vf.R
\name{theme_vf}
\alias{theme_vf}
\title{Theme for 24-2 visual fields plots}
\usage{
theme_vf(base_family = "Helvetica", base_size = 14,
  base_line_size = base_size/22, base_rect_size = base_size/22, ...)
}
\description{
Theme for 24-2 visual fields plots
}
\examples{
p + theme_vf()
}
